requirements.txt
beautifulsoup4==4.13.4
Faker==37.4.0
fastapi==0.115.14
Flask==2.3.1
folium==0.20.0
httpx==0.28.1
numpy==2.3.1
opencv_python==4.10.0.84
Pillow==11.3.0
pillow_avif_plugin==1.5.2
pydantic==2.11.7
python-dotenv==1.1.1
Requests==2.32.4
ultralytics==8.3.162
uvicorn==0.35.0
Werkzeug==3.1.3
